split train param
